Tie Connection.Manager's lifetime to Replication.Supervisor by using another supervisor
